@@ -12,7 +12,7 @@ import android.widget.TextView;
  * @attr ref R.styleable.AutofitTextView_minTextSize
  * @attr ref R.styleable.AutofitTextView_precision
  */
-public class AutofitTextView extends android.support.v7.widget.AppCompatTextView implements AutofitHelper.OnTextSizeChangeListener {
+public class AutofitTextView extends androidx.appcompat.widget.AppCompatTextView implements AutofitHelper.OnTextSizeChangeListener {
 
     private AutofitHelper mHelper;
 
